Add Create and delete wisata

+++ app/Http/Controllers/WisataController.php
@@ -8,10 +8,13 @@ use Illuminate\Support\Str;
 use App\Http\Requests\Wisata as RequestsWisata;
 use App\Models\Galeriwisata;
 use App\Models\Wisata;
+use App\Rules\MaxCountGaleries;
 use Illuminate\Http\Request;
 use Auth;
+use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Validation\Rule;
 
 class WisataController extends Controller
 {
@@ -46,31 +49,42 @@ class WisataController extends Controller
      */
     public function store(RequestsWisata $request, Wisata $wisata)
     {
+        $galeries = request()->file('galeri');
+
+        // Galeries tidak boleh lebih 6 gambar
+        if (count($galeries) > 6) {
+            session()->flash('error', 'Foto tidak boleh lebih dari 6');
+            return redirect(route('wisatas.create'));
+        }
+
+        // WISATAS
         $attr = $request->all();
         $slug = Str::slug($request->nama);
         $attr['slug'] = $slug;
 
-        // Menyimpan File gambar
+        // Menyimpan File gambar wisata
         $gambar = request()->file('gambar');
         $gambarUrl = $gambar->storeAs("images/wisatas", "{$slug}.{$gambar->extension()}");
         $attr['gambar'] = $gambarUrl;
 
+        // Create Data Wisata
         Wisata::create($attr);
 
+        // Get last wisata
+        $lastWisataId = DB::getPdo()->lastInsertId();
 
-        $lastIdWisata = DB::getPdo()->lastInsertId();
-        // GALERY
-        $galeries = request()->file('galeri');
+        // GALERIES
         foreach ($galeries as $galeri) {
-            $rand = rand(0, 10);
+            $rand = rand(0, 100);
             $galeriUrl = $galeri->storeAs("images/galerywisatas", "{$slug}.{$rand}.{$galeri->extension()}");
+            // Create data galeri wisata
             Galeriwisata::create([
                 'galeri' => $galeriUrl,
-                'wisata_id' => $lastIdWisata,
+                'wisata_id' => $lastWisataId,
             ]);
         }
 
-        // Validasi terjadi di RequestWisata (file Request)
+        // Validasi Form in Request File
 
         session()->flash('success', 'Wisata telah ditambahkan');
         return redirect(route('wisatas'));
@@ -87,8 +101,8 @@ class WisataController extends Controller
         if (!$wisata) {
             abort(404);
         }
-
-        return view('wisatas.show', compact('wisata'));
+        $galeriewisatas = DB::table('galeriwisatas')->where('wisata_id', $wisata->id)->get();
+        return view('wisatas.show', compact('wisata', 'galeriewisatas'));
     }
 
     /**
@@ -111,22 +125,51 @@ class WisataController extends Controller
      */
     public function update(Request $request, Wisata $wisata)
     {
+        // Validasi Form Manual
+        $validator = Validator::make($request->all(), [
+            'nama' => [
+                'required',
+                Rule::unique('wisatas')->ignore($wisata->nama),
+            ],
+            'deskripsi' => ['required'],
+            'lokasi' => ['required'],
+            'gambar' => [''],
+            'galeri' => ['required', 'array', 'max:6'], // Ini nanti max 6 buat pengurangan di sama reouest inputnya
+        ]);
+        // dd($validator->fails());
+
+        if ($validator->fails()) {
+            return redirect()->back()->withErrors($validator->errors());
+        }
 
         $attr = $request->all();
         $slug = Str::slug($request->nama);
 
         $attr['slug'] = $slug;
 
-        // Jika tidak ada gambar maka ambil gambar sebelumnya
+        // Update jika request gambar kosong
         $gambar = request()->file('gambar');
         if (isset($gambar)) {
-            Storage::delete($wisata->gambar);
             $gambarUrl = $gambar->storeAs("images/wisatas", "{$slug}.{$gambar->extension()}");
+            Storage::delete($wisata->gambar);
+        } else {
+            $gambarUrl = $wisata->gambar;
+        }
+
+        // Update jika galeri kosong
+        $galeries = request()->galeri;
+        if (isset($galeries)) {
+            foreach ($galeries as $galeri) {
+                Storage::delete($wisata->gambar);
+                $gambarUrl = $gambar->storeAs("images/wisatas", "{$slug}.{$gambar->extension()}");
+            }
         } else {
             $gambarUrl = $wisata->gambar;
         }
 
         $attr['gambar'] = $gambarUrl;
+
+        // Update jika request gambar kosong
 
         // Validasi terjadi di RequestWisata (file Request)
         $wisata->update($attr);
@@ -143,8 +186,22 @@ class WisataController extends Controller
      */
     public function destroy(Wisata $wisata)
     {
-        $wisata->delete();
+        // Delete galeriwisatas Table
+        $galeriewisatas = DB::table('galeriwisatas')->where('wisata_id', $wisata->id)->get();
+        foreach ($galeriewisatas as $galeriwisata) {
+            Storage::delete($galeriwisata->galeri);
+        }
+
+        // Delete galeriwisatas Table
+        DB::table('galeriwisatas')->where('wisata_id', '=', $wisata->id)->delete();
+
+        // Delete wisatas Storage
         Storage::delete($wisata->gambar);
+
+        // Delete wisatas Table
+        $wisata->delete();
+
+
         session()->flash('success', 'Wisata berhasil dihapus');
         return redirect(route('wisatas'));
     }
